Test de Guardar(Ingreso, Gasto) y Leer (Ingreso, Gasto

+++ kakebo/modelos.py
@@ -1,5 +1,7 @@
 from datetime import date
 from enum import Enum
+import csv
+import os
 
 class Movimiento:
     def __init__(self, concepto, fecha, cantidad):
@@ -51,5 +53,26 @@ class CategoriaGastos(Enum):
 class Dao:
     def __init__(self, ruta):
         self.ruta = ruta
-        with open(self.ruta, "w", newline="") as f:
-            f.write("concepto,fecha,cantidad,categoria\n")
+        if not os.path.exists(self.ruta):
+            with open(self.ruta, "w", newline="") as f:
+                f.write("concepto,fecha,cantidad,categoria\n")
+
+    def grabar(self, movimiento):
+        with open(self.ruta, "a", newline="") as f:
+            writer = csv.writer(f, delimiter=",", quotechar='"')
+            if isinstance(movimiento, Ingreso):
+                #f.write(f"{movimiento.concepto},{movimiento.fecha},{movimiento.cantidad},\n") #Sin usar libreria csv
+                writer.writerow([movimiento.concepto,movimiento.fecha,movimiento.cantidad,""])
+            elif isinstance(movimiento, Gasto):
+                writer.writerow([movimiento.concepto,movimiento.fecha,movimiento.cantidad,movimiento.categoria.value])
+                #f.write(f"{movimiento.concepto},{movimiento.fecha},{movimiento.cantidad},{movimiento.categoria.value}\n") #Sin usar libreria csv
+    
+    def leer(self):
+        with open(self.ruta, "r") as f:
+            reader = csv.DictReader(f)
+            for row in reader:
+                if row['categoria'] == "":
+                    variable = Ingreso(row['concepto'], date.fromisoformat(row['fecha']), float(row['cantidad']))
+                else:
+                    variable = Gasto(row["concepto"], date.fromisoformat(row["fecha"]), float(row["cantidad"]), CategoriaGastos(int(row["categoria"])))
+                return variable
